# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Zerodrift Agent API",
    "version": "1.0.0",
    "description": "OpenAPI reference for upload, audit session, workflow, and session group endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque platform token",
        "description": "Use the PLATFORM_TOKEN value as an Authorization: Bearer credential."
      }
    },
    "schemas": {
...
      "UploadTreeEntry": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "File or directory base name."
          },
          "path": {
            "type": "string",
            "description": "Upload-relative path."
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "description": "File size in bytes; omitted for directories."
          },
          "type": {
            "type": "string",
            "enum": [
              "dir",
              "file"
            ],
            "description": "Entry type."
          }
        },
        "required": [
          "name",
          "path",
          "type"
        ]
      },
...
      "UploadTreeResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadTreeEntry"
            },
            "description": "Upload directory entries."
          },
          "path": {
            "type": "string",
            "description": "Requested upload-relative directory path."
          },
          "upload_id": {
            "type": "string",
            "description": "Snake-case upload ID."
          },
          "uploadId": {
            "type": "string",
            "description": "Camel-case upload ID."
          }
        },
        "required": [
          "entries",
          "path",
          "upload_id",
          "uploadId"
        ]
      },
...
      "ZerodriftApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "details": {
            "description": "Optional structured error details."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ZerodriftApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ZerodriftApiError"
              }
            ],
            "description": "Structured ZeroDrift API error."
          },
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Always false for error responses."
          }
        },
        "required": [
          "error",
          "success"
        ]
      }
    },
    "responses": {
...
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                },
                "code": {
                  "type": "string",
                  "example": "401"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied"
                },
                "code": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/zerodrift/uploads/{uploadId}/tree": {
      "get": {
        "operationId": "get-api-zerodrift-uploads-{uploadId}-tree",
        "summary": "List upload tree",
        "description": "Lists files and directories under a stored upload extraction directory.",
        "tags": [
          "ZeroDrift Uploads"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Upload-relative directory path."
            },
            "description": "Upload-relative directory path."
          },
          {
            "in": "path",
            "name": "uploadId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Upload ID."
            },
            "description": "Upload ID.",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Upload directory entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTreeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid upload or path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZerodriftApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
...
}